Move main package to root
