+++ main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/nais/narcos/cmd"
+
+func main() {
+	cmd.Run()
+}
